refactor: split sms.go in multiple files
